Bee can get poisoned by flowers with pesticides lossing health

+++ SaveTheBees/core/src/com/spelunkers/game/screens/EndScreen.java
@@ -30,7 +30,7 @@ public class EndScreen extends ScreenAdapter {
 
     private int beeHivePollen;
     private int beePollen;
-    private int beeHealth;
+    private float beeHealth;
 
     public EndScreen(BeesGame game, Scoreboard scoreboard) {
         this.beeHivePollen = scoreboard.getBeeHivePollenCount();
